// src/model/enums.ts
export enum Tag {
  Low = 'low',
  Medium = 'medium',
  High = 'high',
  Critical = 'critical',
  Design = 'design system',
  Development = 'development',
  Research = 'research',
}

export enum Color {
  Purple = 'purple',
  Yellow = 'yellow',
  Orange = 'orange',
  Red = 'red',
  Pink = 'pink',
  Green = 'green',
  Cyan = 'cyan',
}

export enum Status {
  Todo = 'todo',
  InProgress = 'in_progress',
  Done = 'done',
}